Fixed authentication of logged in user on servlets not managed by spring

+++ kamehouse-testmodule/src/main/java/com/nicobrest/kamehouse/testmodule/servlet/DragonBallUserServlet.java
@@ -74,6 +74,7 @@ public class DragonBallUserServlet extends AbstractKameHouseServlet {
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) {
     try {
+      authorize(request, List.of("ROLE_SAIYAJIN"));
       DragonBallUserDto dragonBallUserDto = getDtoFromRequest(request);
       Long createdId = getDragonBallUserService().create(dragonBallUserDto);
       setResponseBody(response, JsonUtils.toJsonString(createdId));
@@ -86,6 +87,7 @@ public class DragonBallUserServlet extends AbstractKameHouseServlet {
   @Override
   public void doPut(HttpServletRequest request, HttpServletResponse response) {
     try {
+      authorize(request, List.of("ROLE_SAIYAJIN"));
       DragonBallUserDto dragonBallUserDto = getDtoFromRequest(request);
       getDragonBallUserService().update(dragonBallUserDto);
     } catch (KameHouseException e) {
@@ -97,6 +99,7 @@ public class DragonBallUserServlet extends AbstractKameHouseServlet {
   @Override
   public void doDelete(HttpServletRequest request, HttpServletResponse response) {
     try {
+      authorize(request, List.of("ROLE_SAIYAJIN"));
       Long userId = getLongUrlDecodedParam(request, "id");
       DragonBallUser deletedUser = getDragonBallUserService().delete(userId);
       setResponseBody(response, JsonUtils.toJsonString(deletedUser));
@@ -118,5 +121,4 @@ public class DragonBallUserServlet extends AbstractKameHouseServlet {
     dragonBallUserDto.setPowerLevel(Integer.parseInt(getUrlDecodedParam(request, "powerLevel")));
     return dragonBallUserDto;
   }
-
 }
